Proteeti/app/requirements.txt
Flask==3.0.3
flask-dance==7.1.0
requests==2.31.0
python-dotenv==1.0.0
psycopg[binary]>=3.1.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
